requirements.txt
aiohttp==3.9.1
aiohttp_apispec==3.0.0b2
aiohttp_session==2.12.0
alembic==1.13.1
cryptography==42.0.1
asyncpg==0.29.0
greenlet==3.0.3
marshmallow==3.20.2
pytest==7.4.4
PyYAML==6.0.1
pytest-aiohttp==1.0.5
SQLAlchemy==2.0.25
jwt==1.3.1
PyJWT==2.8.0



aiosignal==1.3.1
apispec==6.4.0
attrs==23.2.0
cffi==1.16.0
frozenlist==1.4.1
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
Mako==1.3.0
MarkupSafe==2.1.4
multidict==6.0.4
packaging==23.2
pluggy==1.4.0
pycparser==2.21
pytest-asyncio==0.23.3
typing_extensions==4.9.0
webargs==8.4.0
yarl==1.9.4